Various: mostly 3S-A, MVC2-Rerip, PuzzleFighter
* 3S: Portrait and Stage sprites
* MvC2-Arcade-Rerip: Allow saving
* Most of Puzzle Fighter, just not exposed yet
* SFA2: Rename the Extra 3/4 palettes to Super Trail Light/Dark
* Code cleanup: Move a large bunch of 'int' usage to UINT16

+++ palmod/Game/Game_COTA_A.cpp
@@ -416,7 +416,7 @@ const sDescTreeNode* CGame_COTA_A::GetNodeFromPaletteId(UINT16 nUnitId, UINT16 n
     const sGame_PaletteDataset* paletteSetToUse = nullptr;
     int nDistanceFromZero = nPaletteId;
 
-    for (int nCollectionIndex = 0; nCollectionIndex < nTotalCollections; nCollectionIndex++)
+    for (UINT16 nCollectionIndex = 0; nCollectionIndex < nTotalCollections; nCollectionIndex++)
     {
         const sGame_PaletteDataset* paletteSetToCheck = GetPaletteSet(nUnitId, nCollectionIndex);
         UINT16 nNodeCount;
@@ -466,7 +466,7 @@ const sGame_PaletteDataset* CGame_COTA_A::GetSpecificPalette(UINT16 nUnitId, UIN
     const sGame_PaletteDataset* paletteToUse = nullptr;
     int nDistanceFromZero = nPaletteId;
 
-    for (int nCollectionIndex = 0; nCollectionIndex < nTotalCollections; nCollectionIndex++)
+    for (UINT16 nCollectionIndex = 0; nCollectionIndex < nTotalCollections; nCollectionIndex++)
     {
         const sGame_PaletteDataset* paletteSetToUse = GetPaletteSet(nUnitId, nCollectionIndex);
         UINT16 nNodeCount = GetNodeCountForCollection(nUnitId, nCollectionIndex);
@@ -638,8 +638,8 @@ BOOL CGame_COTA_A::UpdatePalImg(int Node01, int Node02, int Node03, int Node04)
     }
 
     // Default values for multisprite image display for Export
-    int nSrcStart = 0;
-    int nSrcAmt = 0;
+    UINT16 nSrcStart = 0;
+    UINT16 nSrcAmt = 0;
     UINT16 nNodeIncrement = 1;
 
     //Get rid of any palettes if there are any
